working

+++ src/main/java/com/ua/kliushenkov/MudleThree/repository/RaceRepository.java
@@ -3,5 +3,5 @@ package com.ua.kliushenkov.MudleThree.repository;
 import com.ua.kliushenkov.MudleThree.pojo.Race;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RaceRepository extends JpaRepository<Race,Integer> {
+public interface RaceRepository extends JpaRepository<Race,Long> {
 }
